+++ core/docs/changelog.md
@@ -1,4 +1,6 @@
 # TBD
+
+# 1.19.3
 ### Changes
 * Iterate testsuite's tests in alphabetical order